ec-scraper/pyproject.toml
[project]
name = "ec-scraper"
version = "0.1.0"
description = "Intelligent extracurricular opportunity scraper with AI-powered extraction"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.0",
    "crawl4ai>=0.4.0",
    "google-genai>=1.0.0",
    "chromadb>=0.5.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "typer>=0.12.0",
    "aiosqlite>=0.20.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "httpx>=0.27.0",
    "asyncpg>=0.29.0",
    "scrapy>=2.11.0",
    "twisted>=23.10.0",
    "markdownify>=0.11.6",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "aiohttp>=3.9.0",
    "supabase>=2.0.0",
    "numpy>=1.24.0",
    "readability-lxml>=0.8.1",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "ddgs>=7.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[project.scripts]
ec-scraper = "scripts.run_scraper:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
